pyproject.toml
[tool.poetry]
name = "kinetic-models-pipeline"
version = "0.1.0"
description = ""
authors = ["Kian Mehrabani <kianmehrabani@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
PyGithub = "^1.57"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
